+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Add `relative(target, from)` for computing relative paths between URLs.
+
 ## v0.2.0
 
   - Move `Protocol::URL::PATTERN` to `protocol/url/pattern.rb` so it can be shared more easily.